[wip] start adding dynamic routing to case and eligibility screens.
