+++ src/app/(BottomTabNavigation)/AllCases/CaseScreen/[caseUid].tsx
@@ -1,0 +1,66 @@
+import { useLocalSearchParams } from 'expo-router';
+import React, { useEffect, useState } from 'react';
+import { View, ScrollView, Text } from 'react-native';
+
+import styles from './styles';
+import CaseStatusBar from '../../../../Components/CaseStatusBar/CaseStatusBar';
+import CaseSummaryCard from '../../../../Components/CaseSummaryCard/CaseSummaryCard';
+import EducationalBar from '../../../../Components/EducationalBar/EducationalBar';
+import EligibilityCard from '../../../../Components/EligibilityCard/EligibilityCard';
+import FormsCard from '../../../../Components/FormsCard/FormsCard';
+import { getCaseById, getCaseStatus } from '../../../../supabase/queries/cases';
+import { Case, CaseUid, Eligibility } from '../../../../types/types';
+
+function CaseScreen() {
+  const { caseUid } = useLocalSearchParams<{ caseUid: string }>();
+  console.log(caseUid);
+  const [status, setStatus] = useState<Eligibility>();
+  const [caseData, setCaseData] = useState<Case>();
+
+  const getCase = async (uid: string) => {
+    const caseData = await getCaseById(uid);
+    setCaseData(caseData);
+  };
+
+  const getStatus = async (uid: string) => {
+    const caseStatus = await getCaseStatus(uid);
+    setStatus(caseStatus);
+  };
+
+  useEffect(() => {
+    if (caseUid !== undefined) {
+      getCase(caseUid);
+      getStatus(caseUid);
+    }
+  }, []);
+
+  return (
+    <View style={styles.container}>
+      {caseData === undefined ? (
+        <Text>Loading...</Text>
+      ) : (
+        <ScrollView
+          contentContainerStyle={styles.scrollView}
+          showsVerticalScrollIndicator={false}
+        >
+          <View style={styles.titleContainer}>
+            <Text style={styles.title}>{caseData.title}</Text>
+          </View>
+          <CaseStatusBar />
+          {status === Eligibility.ELIGIBLE && (
+            <EligibilityCard caseData={caseData} status={status} />
+          )}
+          <CaseSummaryCard {...caseData} />
+          {(status === Eligibility.INELIGIBLE ||
+            status === Eligibility.UNDETERMINED) && (
+            <EligibilityCard caseData={caseData} status={status} />
+          )}
+          <FormsCard {...caseData} />
+          <EducationalBar />
+        </ScrollView>
+      )}
+    </View>
+  );
+}
+
+export default CaseScreen;
